// src/interfaces/sentTx-response.ts
// Generated by https://quicktype.io

import { Legend } from "./product";

export interface SendTxResponse {
  RESPONSE: Response;
}

export interface Response {
  ID_TX: { text: string };
  NUM_AUTORIZACION: { text: string };
  SALDO: { text: string };
  COMISION: { text: string };
  SALDO_F: { text: string };
  COMISION_F: { text: string };
  FECHA: { text: string };
  MONTO: { text: string };
  MENSAJE: Mensaje;
}

export interface Mensaje {
  CODIGO: { text: string };
  TEXTO: { text: string };
  REFERENCIA: { text: string };
  SALDO?: { text: string };
  PIN?: { text: string };
  legend?: Legend;
}